+++ components/global-navigation-bar/index.jsx
@@ -9,7 +9,10 @@ Neither the name of salesforce.com, inc. nor the names of its contributors may b
 THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 */
 
-// # ContextBar Link Component
+// # Global Navigation Bar Component
+
+// Implements the [Global Navigation Bar design pattern](https://www.lightningdesignsystem.com/components/global-navigation#flavor-navigation-bar) in React.
+// Based on SLDS v2.1.0-dev
 
 // ## Dependencies
 
@@ -19,69 +22,54 @@ import React, { PropTypes } from 'react';
 // ### classNames
 import classNames from 'classnames';
 
-// ### isFunction
-import isFunction from 'lodash.isfunction';
-
-// ### Event Helpers
-import { EventUtil } from '../../utilities';
-
 // ## Constants
-import { CONTEXT_BAR_LINK } from '../../utilities/constants';
+import { GLOBAL_NAVIGATION_BAR } from '../../utilities/constants';
 
 /**
- * Wraps a link in the proper markup to support use in the ContextBar.
+ * Global Navigation Bar represents a list of links that either take the user to another page or parts of the page the user is in.
  */
-const ContextBarLink = (props) => {
-	/* eslint-disable react/prop-types */
-	const {
-		className,
-		label,
-		onClick,
-		...other
-	} = props;
+const GlobalNavigationBar = React.createClass({
+	// ### Display Name
+	// Always use the canonical component name as the React display name.
+	displayName: GLOBAL_NAVIGATION_BAR,
 
-	function handleClick (event) {
-		if (isFunction(onClick)) {
-			EventUtil.trap(event);
+	// ### Prop Types
+	propTypes: {
+		/**
+		 * The items to be displayed in the Global Navigation Bar.
+		 */
+		children: PropTypes.node,
+		/**
+		 * CSS class names to be added to the container element.
+		 */
+		className: PropTypes.oneOfType([PropTypes.array, PropTypes.object, PropTypes.string]),
+		/**
+		 * Typically the cloud name (e.g.- "sales" or "marketing"). This primarily changes the background color.
+		 */
+		cloud: PropTypes.string,
+		/**
+		 * Transforms text and interactions (such as hover) to be more visually accessible.
+		 */
+		theme: PropTypes.oneOf(['light', 'dark'])
+	},
 
-			event.href = this.props.href;
-			onClick(event);
-		}
+	getDefaultProps () {
+		return {
+			cloud: 'default',
+			theme: 'dark'
+		};
+	},
+
+	// ### Render
+	render () {
+		const cloudClass = `slds-context-bar--theme-${this.props.cloud}`;
+		const themeClass = `slds-context-bar--theme-${this.props.theme}`;
+		return (
+			<div className={classNames('slds-context-bar', cloudClass, themeClass, this.props.className)}>
+					{this.props.children}
+			</div>
+		);
 	}
+});
 
-	return (
-		<li className="slds-context-bar__item">
-			<a
-				className={classNames('slds-context-bar__label-action', className)}
-				onClick={handleClick}
-				{...other}
-			>
-				<span className="slds-truncate">{label}</span>
-			</a>
-		</li>
-	);
-};
-
-ContextBarLink.displayName = CONTEXT_BAR_LINK;
-
-// ### Prop Types
-ContextBarLink.propTypes = {
-	/**
-	 * Class names to be added to the anchor element
-	 */
-	className: PropTypes.oneOfType([PropTypes.array, PropTypes.object, PropTypes.string]),
-	/**
-	 * The href of the link.
-	 */
-	href: PropTypes.string,
-	/**
-	 * Text to show for link item.
-	 */
-	label: PropTypes.string,
-	/**
-	 * `function (event)` - fires when the link is clicked. If set, `href` will be ignored, but includes the `href` of the link in the event object.
-	 */
-	onClick: PropTypes.func
-};
-
-module.exports = ContextBarLink;
+module.exports = GlobalNavigationBar;